Instead of using destructor, use "cleanUp()" function
The PHP destructors get called in an inconvenient order. If class B "has
a" instance of class A, when class B is destroyed class A's constructor
is called first, then class B's constructor. If class B needs to take
steps prior to the destruction of class A, it can not.

This commit also changes the INotifyFileMonitor to only accept
registration for one file.

Finally, the INotifyFileMonitor will now gracefully cope with inotify
"ignore" events.

This should have been 3 commits but it got all tangled up as I went down
the rabbit hole.